refactor: add types.ts
